+++ SH6/src/ffof/express/service/IMessageService.java
@@ -9,9 +9,9 @@ public interface IMessageService {
 	
 	public void add(Message msg);
 	public void sendToAll(String postPhone, String content);
-	public void update(int id);
-	public void delete(int id);
-	public void read(int id);
+	public void update(long id);
+	public void delete(long id);
+	public void read(long id);
 	public List<Message> getHistoryMessage(String recvPhone);
 	public Pager<Message> getHistoryMessagePager(String recvPhone, int curPageNum);
 	public List<Message> getUnRead(String recvPhone);
